Add GetModuleList for a given semester function
