+++ uplan-app-api/mocks/modules/get-module-list.json
@@ -1,10 +1,6 @@
 {
   "body": {
-    "moduleId": "8639b267-6c56-44c4-a4dc-c2eb96341783",
     "semesterId": "6beafbd5-9e76-4be0-bf88-52874bc02e3d"
-  },
-  "pathParameters": {
-    "id": "8639b267-6c56-44c4-a4dc-c2eb96341783"
   },
   "requestContext": {
     "identity": {
